9.3.9. Split RoomScreen component
