+++ project/src/components/room-screen/index.ts
@@ -1,4 +1,5 @@
 export {default as OfferInsideItem} from './features-inside';
 export {default as OfferNearsList} from './offer-nears-list';
 export {CommentForm, CommentsList, CommentItem} from './comment';
+export {Gallery, Property, Host} from './internal';
