+++ project/src/components/room-screen/internal/index.ts
@@ -1,0 +1,3 @@
+export {default as Gallery} from './gallery';
+export {default as Property} from './property';
+export {default as Host} from './host';
